Household + SectoralMobility: zero .toDouble, typed friction matrix
Household.scala: all remaining .toDouble eliminated — tryVoluntarySearch,
tryRetraining (Share probability chain), processConsumerCredit (Rate
arithmetic), computeMonthlyFlows (equityIndexReturn: Rate, typed debt
service), updateMpc (Coefficient × Multiplier → Share buffer adjustment),
neighborDistressRatioFast → Share, sectorMobilityRate → Share,
computeAggregates marked @computationBoundary.

SectoralMobility: DefaultFrictionMatrix → Vector[Vector[Share]],
selectTargetSector @computationBoundary (Math.pow), frictionAdjustedParams
fully typed (Share/Multiplier), crossSectorWagePenalty → Multiplier.

SocialConfig: eduRetrainMultiplier/eduWagePremium → Multiplier,
eduSkillRange → (Share, Share).

types.scala: added Rate.toMultiplier.

+++ src/main/scala/com/boombustgroup/amorfati/engine/steps/HouseholdIncomeStep.scala
@@ -54,7 +54,7 @@ object HouseholdIncomeStep:
         depositRates = bsec.banks.map(_ => Banking.hhDepositRate(in.w.nbp.referenceRate)),
       ),
     )
-    val eqReturn           = in.w.financial.equity.monthlyReturn.toDouble
+    val eqReturn           = in.w.financial.equity.monthlyReturn
     val secWages           = if p.flags.sectoralMobility then Some(SectoralMobility.sectorWages(afterWages)) else None
     val secVacancies       =
       if p.flags.sectoralMobility then Some(SectoralMobility.sectorVacancies(afterWages, in.firms)) else None
